ZYAV
Universal playback capabilities, using a new and simple interface, as
well as simple operation, on the Mac for the ultimate continuation of
the ultimate universal user experience service.
Video formats that support playback include, but are not limited to:
mpg, MPEG, MPEG 2, DAT, MP4, m4v, TS, RM, RMVB, WMV, ASF, MKV, avi,
3gp, H264, 3g2, flv, MOV, various HD movies, and various audio files,
and so forth.